feat: handle all version detection in openapi parse function
